+++ arit.h
@@ -1,6 +1,6 @@
 #pragma once
 
-int soma(int a, int b);
-int subtracao(int a, int b);
-int multiplicacao(int a, int b);
-int divisao(int a, int b);
+double soma(double a, double b);
+double subtracao(double a, double b);
+double multiplicacao(double a, double b);
+double divisao(double a, double b);
